Refactoring: Move Options to Options::Main, add Options::Base to be subclass for all options.
