+++ lib/yap/cli/options/addon.rb
@@ -1,59 +1,55 @@
 module Yap
   module Cli
-    class Options::Addon < OptionsBase
-      attr_reader :command, :options
+    module Options
+      class Addon < Base
+        def parse(args)
+          option_parser.order!(args)
+        end
 
-      def initialize(options:)
-        @options = options
-      end
+        def command
+          @command
+        end
 
-      def parse(args)
-        option_parser.order!(args)
-      end
+        def help_message
+          option_parser.to_s
+        end
 
-      def command
-        @command
-      end
+        private
 
-      def help_message
-        option_parser.to_s
-      end
+        def set_command(command)
+          @command = load_command('addon').new
+        end
 
-      private
-
-      def set_command(command)
-        @command = load_command('addon').new
-      end
-
-      def option_parser
-        OptionParser.new do |opts|
-          opts.banner = <<-TEXT.gsub(/^\s*\|/, '')
-            |Usage: #{opts.program_name} addon [command] [addon-name] [options]
-            |
-            |#{Term::ANSIColor.cyan('yap addon')} can be used to control yap addons
-            |
-            |Generate commands:
-            |
-            |  #{Term::ANSIColor.yellow('enable')} - enables a yap addon
-            |  #{Term::ANSIColor.yellow('disable')} - disables a yap addon
-            |
-            |Generate options:
-          TEXT
-
-          opts.on('-h', '--help', 'Prints this help') do
-            puts opts
-            puts
-            puts  <<-TEXT.gsub(/^\s*\|/, '')
-              |Example: disabling an addon
+        def option_parser
+          OptionParser.new do |opts|
+            opts.banner = <<-TEXT.gsub(/^\s*\|/, '')
+              |Usage: #{opts.program_name} addon [command] [addon-name] [options]
               |
-              |   #{opts.program_name} addon disable foo-bar
+              |#{Term::ANSIColor.cyan('yap addon')} can be used to control yap addons
               |
-              |Example: enabling an addon
+              |Generate commands:
               |
-              |   #{opts.program_name} addon enable foo-bar
+              |  #{Term::ANSIColor.yellow('enable')} - enables a yap addon
+              |  #{Term::ANSIColor.yellow('disable')} - disables a yap addon
               |
+              |Generate options:
             TEXT
-            exit 0
+
+            opts.on('-h', '--help', 'Prints this help') do
+              puts opts
+              puts
+              puts  <<-TEXT.gsub(/^\s*\|/, '')
+                |Example: disabling an addon
+                |
+                |   #{opts.program_name} addon disable foo-bar
+                |
+                |Example: enabling an addon
+                |
+                |   #{opts.program_name} addon enable foo-bar
+                |
+              TEXT
+              exit 0
+            end
           end
         end
       end
