+++ lib/yap/cli/options/generate.rb
@@ -1,54 +1,53 @@
 module Yap
   module Cli
-    class Options::Generate < OptionsBase
-      attr_reader :command, :options
+    module Options
+      class Generate < Base
+        def parse(args)
+          if args.empty?
+            args.unshift '--help'
+            STDERR.puts "generate must be called with a component type!"
+            @exit_status = 1
+            puts
+          end
 
-      def initialize(options:)
-        @options = options
-        @exit_status = 0
-      end
-
-      def parse(args)
-        if args.empty?
-          args.unshift '--help'
-          STDERR.puts "generate must be called with a component type!"
-          @exit_status = 1
-          puts
+          option_parser.order!(args)
         end
 
-        option_parser.order!(args)
-      end
+        def command
+          @command ||= load_command('generate').new
+        end
 
-      def command
-        @command ||= load_command('generate').new
-      end
+        def exit_status
+          @exit_status || 0
+        end
 
-      private
+        private
 
-      def option_parser
-        OptionParser.new do |opts|
-          opts.banner = <<-TEXT.gsub(/^\s*\|/, '')
-            |Usage: #{opts.program_name} generate [component-type] [options]
-            |
-            |#{Term::ANSIColor.cyan('yap generate')} can be used to generate yap components, like an addon.
-            |
-            |Generate commands:
-            |
-            |  #{Term::ANSIColor.yellow('addon')} - generates a yap addon skeleton
-            |
-            |Generate options:
-          TEXT
-
-          opts.on('-h', '--help', 'Prints this help') do
-            puts opts
-            puts
-            puts  <<-TEXT.gsub(/^\s*\|/, '')
-              |Example: Generating an addon
+        def option_parser
+          OptionParser.new do |opts|
+            opts.banner = <<-TEXT.gsub(/^\s*\|/, '')
+              |Usage: #{opts.program_name} generate [component-type] [options]
               |
-              |   #{opts.program_name} generate addon magical-key-bindings
+              |#{Term::ANSIColor.cyan('yap generate')} can be used to generate yap components, like an addon.
               |
+              |Generate commands:
+              |
+              |  #{Term::ANSIColor.yellow('addon')} - generates a yap addon skeleton
+              |
+              |Generate options:
             TEXT
-            exit @exit_status
+
+            opts.on('-h', '--help', 'Prints this help') do
+              puts opts
+              puts
+              puts  <<-TEXT.gsub(/^\s*\|/, '')
+                |Example: Generating an addon
+                |
+                |   #{opts.program_name} generate addon magical-key-bindings
+                |
+              TEXT
+              exit exit_status
+            end
           end
         end
       end
